+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "5.4.1"
+__version__ = "5.4.2"
 VERSION = __version__